longitude functions to coords
